Improved usability by automating positioning and scaling of quad billboards
Instead of manually scaling and positioning sprite billboard quads they are now handled by deriving the correct factors in order to maintaining a proper 1:1 world scale of the original model and the billboard sprite. This should make it much easier to configure sprites without having to play around with as many factors to get things working or at least looking correct.