Refactoring for preliminary work on new translation system
-ThreeDeeSpriteRenderer now implements an interface
-Surfaces and SurfaceChains now take an interface for the Renderers
-A new SpriteRendererAdvanced class has been created but is currently exactly the same as the non-advanced one. More to come
-Several constants and logic added to surface and Surface chain.
-Added a new list of 'Dynamic' commands to the Surfaces. Currently unused.

I am working on a new SpriteRenderer that will utilize shaders to perform the 3D models translation into the render target space rather than relying on manually moving the renderer gameobject itself. This will allow for much more flexibility and ease during development as well as retain vital context-sensitive information about the original location of the 3D model. I will be leaving the old system in place however so that it can be used for objects that do not require dynamic information updates regularly or cannot rely on custom shaders to do the translations. More to come soon.